requirements.txt
numpy
pytorch
tensorboard
albumentations
torchvision
requests
